suits/requirements.txt
asgiref==3.11.1
boto3==1.42.58
botocore==1.42.58
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
